refactor: Extract chart config-related action to a separate type
